requirements.txt
# OCR
paddleocr[all]>=3.3.0
paddlepaddle-gpu>=3.3.0

# PDF processing
PyMuPDF>=1.23.0

# Text chunking
langchain>=0.1.0
langchain-text-splitters>=1.0.0

# Vector store
chromadb>=0.3.21

# Embeddings (OpenAI-compatible endpoint)
openai>=1.0.0

# LLM inference (local)
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0

# Evaluation
nltk>=3.8
rouge>=1.0.1
matplotlib>=3.5.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
multiprocess>=0.70.1
python-multipart>=0.0.22

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0
python-dotenv>=1.0.0

# Dev
black>=22.0.0
pylint>=2.0.0
